Have added an implementation of Serializable to all POJO classes (which are receiving from the JSON response). Have added a saving of JSON response for further use as an offline version of app. Have added themes and user's choice may change them.

+++ app/src/main/java/example/yuratoxa/weather/JSONPlaceHolderApi.java
@@ -8,7 +8,7 @@ public interface JSONPlaceHolderApi {
 
     @GET("/data/2.5/forecast")
     Call<WeatherDay> getDataFiveDays(@Query("units") String units, @Query("APPID") String APPID,
-                                     @Query("lang") String lang, @Query("lat") int lat,
-                                     @Query("lon") int lon);
+                                     @Query("lang") String lang, @Query("lat") double lat,
+                                     @Query("lon") double lon);
 }
